added abstract SmartcardDialog class
