+++ common/src/main/java/com/microsoft/identity/common/internal/ui/webview/challengehandlers/SmartcardMaxFailedAttemptsDialog.java
@@ -1,24 +1,20 @@
 package com.microsoft.identity.common.internal.ui.webview.challengehandlers;
 
 import android.app.Activity;
-import android.app.Dialog;
 
 import androidx.appcompat.app.AlertDialog;
 
 import com.microsoft.identity.common.R;
 
 //Show simple dialog when the user has incorrectly attempted to enter their PIN the maximum amount of times allowed.
-public class SmartcardMaxFailedAttemptsDialog {
-
-    private Activity mActivity;
-    private Dialog mDialog;
+public class SmartcardMaxFailedAttemptsDialog extends SmartcardDialog {
 
     public SmartcardMaxFailedAttemptsDialog(Activity activity) {
-        mActivity = activity;
-        CreateDialog();
+        super(activity);
+        createDialog();
     }
 
-    public void CreateDialog() {
+    protected void createDialog() {
         mActivity.runOnUiThread(new Runnable() {
             @Override
             public void run() {
@@ -29,15 +25,6 @@ public class SmartcardMaxFailedAttemptsDialog {
                         .setPositiveButton(R.string.smartcard_max_attempt_dialog_positive_button, null);
                 AlertDialog dialog = builder.create();
                 mDialog = dialog;
-            }
-        });
-    }
-
-    public void show() {
-        mActivity.runOnUiThread(new Runnable() {
-            @Override
-            public void run() {
-                mDialog.show();
             }
         });
     }
